[DEV-11547] Decompose WebBookmarkExtension normalizations
