+++ packages/slate-editor/src/extensions/web-bookmark/lib/index.ts
@@ -1,3 +1,2 @@
 export { createWebBookmark } from './createWebBookmark';
-export { normalizeRedundantWebBookmarkAttributes, normalizeUrlAttribute } from './normalizations';
 export { parseSerializedElement } from './parseSerializedElement';
