+++ packages/slate-editor/src/extensions/web-bookmark/normalizations/index.ts
@@ -1,0 +1,2 @@
+export { unsetUnknownAttributes } from './unsetUnknownAttributes';
+export { normalizeUrlAttribute } from './normalizeUrlAttribute';
